Implement Covariance Storage, Tracking, Interpolation (#59)
# Pull Request

## Description

This PR adds the ability to store covariances in `OrbitTrajectory`. It
also defines the `CovarianceProvider` trait which implements similar
access patterns as `StateProvider`. To enable parity of behavior we
implement matrix-interpolation through a simple matrix square root as
well as a two-wasserstein metric.

## Changelog

### Added
- Implement `CovarianceProvider` trait for `OrbitTrajectory`. Provides
`covariance`, `covariance_eci`, `covariance_gcrf`, and `covariance_rtn`
- Add `interpolation` submodule to store consistent covariance
interpolation methods
- Implement covariance rotation from ECI to RTN frame
- Extends `OrbitTrajectory` to optionally store covariance information.